refactor: move oracle_node object into its own file
