+++ oracle/node/tests/functions/get_price.rs
@@ -1,0 +1,9 @@
+
+mod success {
+    use super::*;
+
+    #[tokio::test]
+    async fn can_get_price() {
+        
+    }
+}
