+++ oracle/node/tests/functions/mod.rs
@@ -1,0 +1,3 @@
+mod get_price;
+mod new;
+mod run;
