+++ oracle/node/tests/functions/new.rs
@@ -1,0 +1,9 @@
+
+mod success {
+    use super::*;
+
+    #[tokio::test]
+    async fn can_initialize_oracle_node() {
+        
+    }
+}
