+++ oracle/node/tests/functions/run.rs
@@ -1,0 +1,9 @@
+
+mod success {
+    use super::*;
+
+    #[tokio::test]
+    async fn can_run() {
+    
+    }
+}
